renderer: widen RenderTok.length int16 -> int32
int16(s.len) overflowed (RangeDefect) on string literals larger than 32 KB —
e.g. generated data tables. The field is write-only within the renderer, so
widening is safe.

+++ src/nimony/renderer.nim
@@ -72,7 +72,7 @@ type
   RenderFlags* = set[RenderFlag]
   RenderTok* = object
     kind*: TokType
-    length*: int16
+    length*: int32   # was int16 — overflowed on string literals > 32 KB
     sym*: SymId
     isDef*: bool
 
@@ -132,7 +132,7 @@ proc initSrcGen(renderFlags: RenderFlags): SrcGen =
 
 proc addTok(g: var SrcGen, kind: TokType, s: string; sym: SymId = SymId(0);
             isDef = false) =
-  g.tokens.add RenderTok(kind: kind, length: int16(s.len), sym: sym, isDef: isDef)
+  g.tokens.add RenderTok(kind: kind, length: int32(s.len), sym: sym, isDef: isDef)
   g.buf.add(s)
   if kind != tkSpaces:
     inc g.col, s.len
